Rearranged Properties
Since I added multiple classes to the html file, here I was now able to put only the font-size on ".big-red-text". This made the code a little more streamlined. I also just rearranged the order to have the class and id selectors next to each other. The changes made to the html file that relates to the change in this css file can be found on my html commit titled "Made Changes Commented on in Previous Commit".